requirement.txt
psycopg[binary]==3.1.12
gunicorn==21.2.0
django-cors-headers==4.3.1
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-ninja==1.1.0
idna==3.6
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
whitenoise